CRI-22 feat: init moved shadcn-components
