--- apps/web-client/src/shared/libs/cn/index.ts
@@ -1,1 +1,0 @@
-export { cn } from './cn.ts';
