+++ apps/web-client/src/shared/libs/index.ts
@@ -1,1 +1,1 @@
-export * from "./cn"
+export * from "../../../../../packages/ui/src/libs"
