--- apps/web-client/src/shared/ui/shadcn-components/button/index.ts
@@ -1,1 +1,0 @@
-export { Button } from './button.tsx';
