--- apps/web-client/src/shared/ui/shadcn-components/card/index.ts
@@ -1,1 +1,0 @@
-export { Card, CardAction, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from './card.tsx';
