--- apps/web-client/src/shared/ui/shadcn-components/input/index.ts
@@ -1,1 +1,0 @@
-export { Input } from './input.tsx';
