--- apps/web-client/src/shared/ui/shadcn-components/switch/index.ts
@@ -1,1 +1,0 @@
-export { Switch } from './switch.tsx';
