--- apps/web-client/tailwind.config.js
@@ -1,3 +1,0 @@
-export default {
-  content: ["./src/**/*.{js,ts,jsx,tsx}"],
-}
