+++ packages/ui/components.json
@@ -5,13 +5,13 @@
   "tsx": true,
   "tailwind": {
     "config": "",
-    "css": "../../packages/ui/build/src/styles/index.css",
+    "css": "./build/src/styles/index.css",
     "baseColor": "zinc",
     "cssVariables": true,
     "prefix": ""
   },
   "aliases": {
-    "components": "@bc-arch-drafter/ui/shadcn-components",
+    "components": "@/shadcn-components",
     "utils": "@/lib/utils",
     "lib": "@/lib"
   },
