+++ packages/ui/package.json
@@ -1,10 +1,10 @@
 {
   "name": "@bc-arch-drafter/ui",
-  "version": "0.0.1",
+  "version": "0.0.6",
   "type": "module",
   "scripts": {
     "clean": "rm -rf build",
-    "build": "tsc && tsc-alias",
+    "build": "tsc && tsc-alias && copyfiles src/styles/*.css build/src/styles",
     "build:watch": "concurrently \"tsc -w\" \"tsc-alias -w\"",
     "lint": "eslint --cache ."
   },
@@ -12,5 +12,25 @@
     "types": "./build/src/index.d.ts",
     "default": "./build/src/index.js"
   },
-  "dependencies": {}
+  "style": "build/src/styles/index.css",
+  "dependencies": {
+    "@radix-ui/react-checkbox": "^1.3.2",
+    "@radix-ui/react-slot": "^1.2.3",
+    "@radix-ui/react-switch": "^1.2.5",
+    "@tailwindcss/vite": "^4.1.11",
+    "@bc-arch-drafter/ui": "workspace:*",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "lucide-react": "^0.534.0",
+    "tailwind-merge": "^3.3.1",
+    "tailwindcss": "^4.1.11",
+    "tw-animate-css": "^1.3.6"
+  },
+  "peerDependencies": {
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0"
+  },
+  "devDependencies": {
+    "copyfiles": "^2.4.1"
+  }
 }
