+++ packages/ui/src/shadcn-components/card/card.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '@/shared/libs';
+import { cn } from '@/lib/cn';
 
 function Card({ className, ...props }: React.ComponentProps<'div'>) {
   return (
