+++ packages/ui/src/shadcn-components/card/index.ts
@@ -1,0 +1,7 @@
+export { Card,
+  CardAction,
+  CardContent,
+  CardDescription,
+  CardFooter,
+  CardHeader,
+  CardTitle } from './card';
