+++ packages/ui/src/shadcn-components/checkbox/checkbox.tsx
@@ -2,7 +2,7 @@ import * as CheckboxPrimitive from "@radix-ui/react-checkbox"
 import { CheckIcon } from "lucide-react"
 import * as React from "react"
 
-import { cn } from '@/shared/libs';
+import { cn } from '@/lib/cn';
 
 function Checkbox({
   className,
