+++ packages/ui/src/shadcn-components/switch/index.ts
@@ -1,0 +1,1 @@
+export { Switch } from './switch';
