+++ packages/ui/src/styles.d.ts
@@ -1,0 +1,4 @@
+declare module '*.css' {
+  const content: string;
+  export default content;
+}
