+++ tsconfig.base.json
@@ -1,6 +1,8 @@
 {
   "compilerOptions": {
-    "lib": ["esnext"],
+    "lib": [
+      "esnext"
+    ],
     "module": "esnext",
     "esModuleInterop": true,
     "resolveJsonModule": true,
@@ -13,6 +15,14 @@
     "incremental": true,
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
+    "paths": {
+      "@bc-arch-drafter/*": [
+        "packages/*/src"
+      ]
+    }
   },
-  "exclude": ["node_modules", "*/build/**"]
+  "exclude": [
+    "node_modules",
+    "*/build/**"
+  ]
 }
